remove all remaining palettes that match

+++ delete_palettes.sh
@@ -5,16 +5,18 @@ palfiles=$(find graphics -type f -name '*.gbapal')
 for f in $palfiles
 do
     base=$(echo $f | cut -f 1 -d '.')
-    if [ -e "$base.png" ]
+    if [ -e "$base.png" ] && [ -e "$base.pal" ]
     then
-        tools/gbagfx/gbagfx "$base.png" "$base.new.gbapal"
-        cmp "$f" "$base.new.gbapal"
-        if [ $? ]
+        if tools/gbagfx/gbagfx "$base.png" "$base.new.gbapal"
         then
-            echo "warning: $base.png contains a nonmatching palette"
-        else
-            rm "$base.pal"
+            if cmp --silent "$f" "$base.new.gbapal"
+            then
+                rm "$base.pal"
+                #echo "deleted $base.pal"
+            else
+                echo "warning: $base.png contains a nonmatching palette"
+            fi
+            rm "$base.new.gbapal"
         fi
-        rm "$base.new.gbapal"
     fi
 done
